Move CI from Travis to Github Actions
- Remove the 'dev_precommit' extra, because all other dependencies
  are anyway required to run pylint, so 'dev' can be used
- Move the 'test script' content into the CI configuration
- Fix some cyclic import warnings
- Add a documentation build CI job
