+++ .ci/install_script.sh
@@ -5,10 +5,7 @@
 # Be verbose, and stop with error as soon there's one
 set -ev
 
-pip install -U 'pip<19' wheel setuptools
-pip install git+https://github.com/greschd/aiida-wannier90.git@edge
-
-cd ${TRAVIS_BUILD_DIR}
+pip install -U pip wheel setuptools
 
 case "$INSTALL_TYPE" in
     testing)
@@ -19,7 +16,10 @@ case "$INSTALL_TYPE" in
         ls -1 dist/ | xargs -I % pip install dist/%[testing,strain]
         ;;
     dev_precommit)
-        pip install .[dev_precommit,strain]
+        pip install .[dev]
+        ;;
+    docs)
+        pip install .[docs]
         ;;
 esac
