+++ .github/workflows/ci.yml
@@ -1,0 +1,100 @@
+name: Continuous Integration
+
+on: [push, pull_request]
+
+jobs:
+  docs:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v1
+      - uses: actions/cache@v2
+        with:
+          path: ~/.cache/pip
+          key: pip-docs-${{ hashFiles('**/setup.json') }}
+          restore-keys: pip-docs-
+      - name: Set up Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.8
+      - name: Install python dependencies
+        env:
+          INSTALL_TYPE: docs
+        run: .ci/install_script.sh
+      - name: Build documentation
+        env:
+          READTHEDOCS: "True"
+        run: SPHINXOPTS='-nW' make -C doc html
+      - uses: actions/upload-artifact@v2
+        with:
+          name: doc-build
+          path: doc/build/html
+
+  pre-commit:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Cache python dependencies
+        id: cache-pip
+        uses: actions/cache@v1
+        with:
+          path: ~/.cache/pip
+          key: pip-pre-commit-${{ hashFiles('**/setup.json') }}
+          restore-keys: pip-pre-commit-
+      - name: Set up Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: 3.8
+      - name: Install python dependencies
+        env:
+          INSTALL_TYPE: dev_precommit
+        run: .ci/install_script.sh
+      - name: Run pre-commit
+        run: pre-commit run --all-files || ( git status --short ; git diff ; exit 1 )
+
+  tests:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: [3.6, 3.8]
+        install-type: [testing]
+        include:
+          - python-version: 3.8
+            install-type: testing_sdist
+    services:
+      postgres:
+        image: postgres:10
+        ports:
+          - 5432:5432
+      rabbitmq:
+        image: rabbitmq:latest
+        ports:
+          - 5672:5672
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v2
+      - name: Cache python dependencies
+        id: cache-pip
+        uses: actions/cache@v1
+        with:
+          path: ~/.cache/pip
+          key: pip-${{ matrix.python-version }}-tests-${{ hashFiles('**/setup.json') }}
+          restore-keys: pip-${{ matrix.python-version }}-tests
+      - name: Set up Python ${{ matrix.python-version }}
+        uses: actions/setup-python@v2
+        with:
+          python-version: ${{ matrix.python-version }}
+      - name: Install python dependencies
+        env:
+          INSTALL_TYPE: ${{ matrix.install-type }}
+        run: .ci/install_script.sh
+      - name: Install Wannier90
+        run: ./build_wannier90.sh
+        working-directory: .ci
+      - name: Set up test configuration
+        run: |
+          python configure.py . test_config.yml ../tests/config.yml
+          cat ../tests/config.yml
+        working-directory: .ci
+      - name: Run pytest
+        run: pytest --skip-qe --skip-vasp --quiet-wipe --print-status
+        working-directory: tests
