--- .travis.yml
@@ -1,31 +1,0 @@
-language: python
-dist: bionic
-python:
-  - "3.6"
-  - "3.7"
-  - "3.8"
-env:
-  - TEST_TYPE="tests" INSTALL_TYPE="testing"
-  - TEST_TYPE="tests" INSTALL_TYPE="testing_sdist"
-jobs:
-  include:
-  - python: "3.7"
-    env: TEST_TYPE="pre-commit" INSTALL_TYPE="dev_precommit"
-cache: pip
-sudo: true
-services:
-  - postgresql
-  - rabbitmq
-addons:
-  apt:
-    packages:
-      - rabbitmq-server
-      - gfortran
-      - libblas-dev
-      - liblapack-dev
-before_install:
-  - sudo service postgresql restart
-install:
-  - ./.travis-data/install_script.sh
-script:
-  - travis_wait 60 ./.travis-data/test_script.sh
